fetch/metadata/generated/element-script-sec-fetch-site.https.sub.html

<!DOCTYPE html>
<!--
This test was procedurally generated. Please do not modify it directly.
Sources:
- fetch/metadata/tools/cases/master.yml
- fetch/metadata/tools/templates/element-script.sub.html
-->
<html lang="en">
  <meta charset="utf-8">
  <title>Fetch Metadata - Sec-Fetch-Site - redirection from HTTPS</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script src="/fetch/metadata/resources/helper.sub.js"></script>
  <body>
  <script>
  'use strict';

  function induceRequest(url, attributes) {
    const script = document.createElement('script');
    script.setAttribute('src', url);

    for (const [ name, value ] of Object.entries(attributes)) {
      script.setAttribute(name, value);
    }

    return new Promise((resolve, reject) => {
        script.onload = resolve;
        script.onerror = () => reject('Failed to load script');
        document.body.appendChild(script);
      })
      .then(() => script.remove());
  }

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameOrigin', 'crossSite', 'sameOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Same-Origin -> Cross-Site -> Same-Origin redirect, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameOrigin', 'crossSite', 'sameOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {"type": "module"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Same-Origin -> Cross-Site -> Same-Origin redirect, attributes: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['crossSite', 'sameOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Cross-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['crossSite', 'sameOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {"type": "module"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Cross-Site -> Same Origin, attributes: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['crossSite', 'sameSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Cross-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['crossSite', 'sameSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {"type": "module"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Cross-Site -> Same-Site, attributes: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['crossSite', 'crossSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Cross-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['crossSite', 'crossSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {"type": "module"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Cross-Site -> Cross-Site, attributes: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameOrigin', 'sameOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'Same-Origin -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameOrigin', 'sameOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {"type": "module"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-origin');
        });
  }, 'Same-Origin -> Same Origin, attributes: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameOrigin', 'sameSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'Same-Origin -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameOrigin', 'sameSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {"type": "module"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'Same-Origin -> Same-Site, attributes: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameOrigin', 'crossSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Same-Origin -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameOrigin', 'crossSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {"type": "module"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Same-Origin -> Cross-Site, attributes: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameSite', 'sameOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'Same-Site -> Same Origin, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameSite', 'sameOrigin'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {"type": "module"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'Same-Site -> Same Origin, attributes: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameSite', 'sameSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'Same-Site -> Same-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameSite', 'sameSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {"type": "module"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'same-site');
        });
  }, 'Same-Site -> Same-Site, attributes: type=module');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameSite', 'crossSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Same-Site -> Cross-Site, no attributes');

  promise_test(() => {
    const key = '{{uuid()}}';
    const url = makeRequestURL(
      key, ['sameSite', 'crossSite'], { mime: 'application/javascript' }
    );

    return induceRequest(url,
        {"type": "module"}
      )
      .then(() => retrieve(key))
      .then((headers) => {
          assert_own_property(headers, 'sec-fetch-site');
          assert_equals(headers['sec-fetch-site'], 'cross-site');
        });
  }, 'Same-Site -> Cross-Site, attributes: type=module');
  </script>
</html>
